chat_processor/requirements.txt
fastapi==0.120.0
python-dotenv==1.1.1
uvicorn==0.38.0